Adding user factory
