+++ apps/ex_marketer/test/support/factory.ex
@@ -1,5 +1,6 @@
 defmodule ExMarketer.Factory do
   use ExMachina.Ecto, repo: ExMarketer.Repo
 
+  use ExMarketer.UserFactory
   use ExMarketer.KeywordFactory
 end
