+++ apps/ex_marketer_web/test/ex_marketer_web/controllers/accounts/user_registration_controller_test.exs
@@ -1,8 +1,6 @@
 defmodule ExMarketerWeb.Accounts.UserRegistrationControllerTest do
   use ExMarketerWeb.ConnCase, async: true
 
-  import ExMarketer.AccountsFactory
-
   describe "GET /users/register" do
     test "renders registration page", %{conn: conn} do
       conn = get(conn, Routes.user_registration_path(conn, :new))
@@ -13,7 +11,7 @@ defmodule ExMarketerWeb.Accounts.UserRegistrationControllerTest do
     end
 
     test "redirects if already logged in", %{conn: conn} do
-      conn = conn |> log_in_user(user_fixture()) |> get(Routes.user_registration_path(conn, :new))
+      conn = conn |> log_in_user(insert(:user)) |> get(Routes.user_registration_path(conn, :new))
       assert redirected_to(conn) == "/"
     end
   end
