+++ apps/ex_marketer_web/test/ex_marketer_web/controllers/accounts/user_reset_password_controller_test.exs
@@ -3,10 +3,9 @@ defmodule ExMarketerWeb.Accounts.UserResetPasswordControllerTest do
 
   alias ExMarketer.Accounts
   alias ExMarketer.Repo
-  import ExMarketer.AccountsFactory
 
   setup do
-    %{user: user_fixture()}
+    %{user: insert(:user)}
   end
 
   describe "GET /users/reset_password" do
