+++ apps/ex_marketer_web/test/ex_marketer_web/controllers/accounts/user_settings_controller_test.exs
@@ -2,7 +2,6 @@ defmodule ExMarketerWeb.Accounts.UserSettingsControllerTest do
   use ExMarketerWeb.ConnCase, async: true
 
   alias ExMarketer.Accounts
-  import ExMarketer.AccountsFactory
 
   setup :register_and_log_in_user
