+++ apps/ex_marketer_web/test/ex_marketer_web/features/accounts/login_test.exs
@@ -1,13 +1,11 @@
 defmodule ExMarketerWeb.Accounts.LoginTest do
   use ExMarketerWeb.FeatureCase, async: true
 
-  alias ExMarketer.AccountsFactory
-
   feature "Login", %{session: session} do
-    user = AccountsFactory.user_fixture()
+    user = insert(:user)
 
     session
-    |> login(user.email, AccountsFactory.valid_user_password())
+    |> login(user.email, valid_user_password())
     |> assert_has(Query.css(".app-layout"))
     |> assert_has(Query.text(user.email))
   end
