+++ apps/ex_marketer_web/test/ex_marketer_web/features/keyword/view_list_keyword_test.exs
@@ -1,8 +1,6 @@
 defmodule ExMarketerWeb.ViewListKeywordTest do
   use ExMarketerWeb.FeatureCase, async: true
 
-  alias ExMarketer.AccountsFactory
-
   @path Routes.keyword_path(ExMarketerWeb.Endpoint, :index)
 
   @messages %{
@@ -18,14 +16,14 @@ defmodule ExMarketerWeb.ViewListKeywordTest do
 
   describe "given the keyword in the database" do
     feature "view list keyword", %{session: session} do
-      user = AccountsFactory.user_fixture()
+      user = insert(:user)
       insert(:keyword, status: "created")
       insert(:keyword, status: "in_progress")
       insert(:keyword, status: "failed")
       insert(:keyword, status: "successed")
 
       session
-      |> login(user.email, AccountsFactory.valid_user_password())
+      |> login(user.email, valid_user_password())
       |> visit(@path)
       |> refute_has(Query.text(@messages.empty_data))
       |> assert_has(Query.css(@selectors.card_keyword, count: 4))
@@ -41,10 +39,10 @@ defmodule ExMarketerWeb.ViewListKeywordTest do
 
   describe "given NO keyword in the database" do
     feature "view list keyword", %{session: session} do
-      user = AccountsFactory.user_fixture()
+      user = insert(:user)
 
       session
-      |> login(user.email, AccountsFactory.valid_user_password())
+      |> login(user.email, valid_user_password())
       |> visit(@path)
       |> assert_has(Query.text(@messages.empty_data))
       |> refute_has(Query.css(@selectors.card_keyword))
