+++ apps/ex_marketer_web/test/support/conn_case.ex
@@ -30,6 +30,32 @@ defmodule ExMarketerWeb.ConnCase do
 
       # The default endpoint for testing
       @endpoint ExMarketerWeb.Endpoint
+
+      @doc """
+      Setup helper that registers and logs in users.
+
+          setup :register_and_log_in_user
+
+      It stores an updated connection and a registered user in the
+      test context.
+      """
+      def register_and_log_in_user(%{conn: conn}) do
+        user = insert(:user)
+        %{conn: log_in_user(conn, user), user: user}
+      end
+
+      @doc """
+      Logs the given `user` into the `conn`.
+
+      It returns an updated `conn`.
+      """
+      def log_in_user(conn, user) do
+        token = ExMarketer.Accounts.generate_user_session_token(user)
+
+        conn
+        |> Phoenix.ConnTest.init_test_session(%{})
+        |> Plug.Conn.put_session(:user_token, token)
+      end
     end
   end
 
@@ -41,31 +67,5 @@ defmodule ExMarketerWeb.ConnCase do
     end
 
     {:ok, conn: Phoenix.ConnTest.build_conn()}
-  end
-
-  @doc """
-  Setup helper that registers and logs in users.
-
-      setup :register_and_log_in_user
-
-  It stores an updated connection and a registered user in the
-  test context.
-  """
-  def register_and_log_in_user(%{conn: conn}) do
-    user = ExMarketer.AccountsFactory.user_fixture()
-    %{conn: log_in_user(conn, user), user: user}
-  end
-
-  @doc """
-  Logs the given `user` into the `conn`.
-
-  It returns an updated `conn`.
-  """
-  def log_in_user(conn, user) do
-    token = ExMarketer.Accounts.generate_user_session_token(user)
-
-    conn
-    |> Phoenix.ConnTest.init_test_session(%{})
-    |> Plug.Conn.put_session(:user_token, token)
   end
 end
