+++ apps/ex_marketer_web/test/support/factory.ex
@@ -1,5 +1,6 @@
 defmodule ExMarketerWeb.Factory do
   use ExMachina.Ecto, repo: ExMarketer.Repo
 
+  use ExMarketer.UserFactory
   use ExMarketer.KeywordFactory
 end
